gh-135953: Refactor test_sampling_profiler into multiple files
