+++ Lib/test/test_profiling/test_sampling_profiler/__init__.py
@@ -1,0 +1,9 @@
+"""Tests for the sampling profiler (profiling.sampling)."""
+
+import os
+from test.support import load_package_tests
+
+
+def load_tests(*args):
+    """Load all tests from this subpackage."""
+    return load_package_tests(os.path.dirname(__file__), *args)
